+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Icons`: Add `GoogleIconNext` and `FacebookIconColorNext` components.
+
 ## [10.9.0] - 2022-05-17
 
 Features: